moved gc test to conftest
